+++ crates/oxc_semantic/tests/main.rs
@@ -16,11 +16,10 @@ fn get_scope_snapshot(semantic: &Semantic, scopes: impl Iterator<Item = ScopeId>
         }
         let flags = semantic.scopes().get_flags(scope_id);
         result.push('{');
-        if let Some(child_ids) = semantic.scopes().get_child_ids(scope_id) {
-            result.push_str("\"children\":");
-            result.push_str(&get_scope_snapshot(semantic, child_ids.iter().copied()));
-            result.push(',');
-        }
+        let child_ids = semantic.scopes().get_child_ids(scope_id);
+        result.push_str("\"children\":");
+        result.push_str(&get_scope_snapshot(semantic, child_ids.iter().copied()));
+        result.push(',');
         result.push_str(format!("\"flags\": \"{flags:?}\",").as_str());
         result.push_str(format!("\"id\": {},", scope_id.index()).as_str());
         result.push_str(