+++ crates/oxc_traverse/src/context/scoping.rs
@@ -2,10 +2,7 @@ use std::{cell::Cell, str};
 
 use compact_str::{format_compact, CompactString};
 #[allow(clippy::wildcard_imports)]
-use oxc_ast::{
-    ast::*,
-    visit::{walk, Visit},
-};
+use oxc_ast::{ast::*, visit::Visit};
 use oxc_semantic::{AstNodeId, Reference, ScopeTree, SymbolTable};
 use oxc_span::{Atom, CompactStr, Span, SPAN};
 use oxc_syntax::{
@@ -122,10 +119,8 @@ impl TraverseScoping {
 
     fn insert_scope_below(&mut self, child_scope_ids: &[ScopeId], flags: ScopeFlags) -> ScopeId {
         // Remove these scopes from parent's children
-        if let Some(current_child_scope_ids) = self.scopes.get_child_ids_mut(self.current_scope_id)
-        {
-            current_child_scope_ids.retain(|scope_id| !child_scope_ids.contains(scope_id));
-        }
+        let current_child_scope_ids = self.scopes.get_child_ids_mut(self.current_scope_id);
+        current_child_scope_ids.retain(|scope_id| !child_scope_ids.contains(scope_id));
 
         // Create new scope as child of parent
         let new_scope_id = self.create_child_scope_of_current(flags);
@@ -286,12 +281,7 @@ impl TraverseScoping {
         flags: ReferenceFlags,
     ) -> IdentifierReference<'a> {
         let reference_id = self.create_bound_reference(symbol_id, flags);
-        IdentifierReference {
-            span,
-            name,
-            reference_id: Cell::new(Some(reference_id)),
-            reference_flags: flags,
-        }
+        IdentifierReference { span, name, reference_id: Cell::new(Some(reference_id)) }
     }
 
     /// Create an unbound reference
@@ -314,12 +304,7 @@ impl TraverseScoping {
         flags: ReferenceFlags,
     ) -> IdentifierReference<'a> {
         let reference_id = self.create_unbound_reference(name.to_compact_str(), flags);
-        IdentifierReference {
-            span,
-            name,
-            reference_id: Cell::new(Some(reference_id)),
-            reference_flags: flags,
-        }
+        IdentifierReference { span, name, reference_id: Cell::new(Some(reference_id)) }
     }
 
     /// Create a reference optionally bound to a `SymbolId`.
@@ -383,6 +368,35 @@ impl TraverseScoping {
             }));
         let symbol_id = reference.symbol_id();
         self.create_reference_id(ident.span, ident.name.clone(), symbol_id, flags)
+    }
+
+    /// Determine whether evaluating the specific input `node` is a consequenceless reference.
+    ///
+    /// I.E evaluating it won't result in potentially arbitrary code from being ran. The following are
+    /// allowed and determined not to cause side effects:
+    ///
+    /// - `this` expressions
+    /// - `super` expressions
+    /// - Bound identifiers
+    ///
+    /// Based on Babel's `scope.isStatic` logic.
+    /// <https://github.com/babel/babel/blob/419644f27c5c59deb19e71aaabd417a3bc5483ca/packages/babel-traverse/src/scope/index.ts#L557>
+    ///
+    /// # Panics
+    /// Can only panic if [`IdentifierReference`] does not have a reference_id, which it always should.
+    #[inline]
+    pub fn is_static(&self, expr: &Expression) -> bool {
+        match expr {
+            Expression::ThisExpression(_) | Expression::Super(_) => true,
+            Expression::Identifier(ident) => self
+                .symbols
+                .get_reference(ident.reference_id.get().unwrap())
+                .symbol_id()
+                .is_some_and(|symbol_id| {
+                    self.symbols.get_resolved_references(symbol_id).all(|r| !r.is_write())
+                }),
+            _ => false,
+        }
     }
 }
 
@@ -514,86 +528,90 @@ impl ChildScopeCollector {
     fn new() -> Self {
         Self { scope_ids: vec![] }
     }
+
+    fn add_scope(&mut self, scope_id: &Cell<Option<ScopeId>>) {
+        self.scope_ids.push(scope_id.get().unwrap());
+    }
 }
 
 impl<'a> Visit<'a> for ChildScopeCollector {
+    #[inline]
     fn visit_block_statement(&mut self, stmt: &BlockStatement<'a>) {
-        self.scope_ids.push(stmt.scope_id.get().unwrap());
+        self.add_scope(&stmt.scope_id);
     }
 
+    #[inline]
     fn visit_for_statement(&mut self, stmt: &ForStatement<'a>) {
-        if let Some(scope_id) = stmt.scope_id.get() {
-            self.scope_ids.push(scope_id);
-        } else {
-            walk::walk_for_statement(self, stmt);
-        }
+        self.add_scope(&stmt.scope_id);
     }
 
+    #[inline]
     fn visit_for_in_statement(&mut self, stmt: &ForInStatement<'a>) {
-        if let Some(scope_id) = stmt.scope_id.get() {
-            self.scope_ids.push(scope_id);
-        } else {
-            walk::walk_for_in_statement(self, stmt);
-        }
+        self.add_scope(&stmt.scope_id);
     }
 
+    #[inline]
     fn visit_for_of_statement(&mut self, stmt: &ForOfStatement<'a>) {
-        if let Some(scope_id) = stmt.scope_id.get() {
-            self.scope_ids.push(scope_id);
-        } else {
-            walk::walk_for_of_statement(self, stmt);
-        }
+        self.add_scope(&stmt.scope_id);
     }
 
+    #[inline]
     fn visit_switch_statement(&mut self, stmt: &SwitchStatement<'a>) {
-        self.scope_ids.push(stmt.scope_id.get().unwrap());
+        self.add_scope(&stmt.scope_id);
     }
 
+    #[inline]
     fn visit_catch_clause(&mut self, clause: &CatchClause<'a>) {
-        self.scope_ids.push(clause.scope_id.get().unwrap());
+        self.add_scope(&clause.scope_id);
     }
 
-    fn visit_finally_clause(&mut self, clause: &BlockStatement<'a>) {
-        self.scope_ids.push(clause.scope_id.get().unwrap());
+    #[inline]
+    fn visit_finally_clause(&mut self, block: &BlockStatement<'a>) {
+        self.add_scope(&block.scope_id);
     }
 
+    #[inline]
     fn visit_function(&mut self, func: &Function<'a>, _flags: ScopeFlags) {
-        self.scope_ids.push(func.scope_id.get().unwrap());
+        self.add_scope(&func.scope_id);
     }
 
+    #[inline]
     fn visit_class(&mut self, class: &Class<'a>) {
-        if let Some(scope_id) = class.scope_id.get() {
-            self.scope_ids.push(scope_id);
-        } else {
-            walk::walk_class(self, class);
-        }
+        self.add_scope(&class.scope_id);
     }
 
+    #[inline]
     fn visit_static_block(&mut self, block: &StaticBlock<'a>) {
-        self.scope_ids.push(block.scope_id.get().unwrap());
+        self.add_scope(&block.scope_id);
     }
 
+    #[inline]
     fn visit_arrow_function_expression(&mut self, expr: &ArrowFunctionExpression<'a>) {
-        self.scope_ids.push(expr.scope_id.get().unwrap());
+        self.add_scope(&expr.scope_id);
     }
 
+    #[inline]
     fn visit_ts_enum_declaration(&mut self, decl: &TSEnumDeclaration<'a>) {
-        self.scope_ids.push(decl.scope_id.get().unwrap());
+        self.add_scope(&decl.scope_id);
     }
 
+    #[inline]
     fn visit_ts_module_declaration(&mut self, decl: &TSModuleDeclaration<'a>) {
-        self.scope_ids.push(decl.scope_id.get().unwrap());
+        self.add_scope(&decl.scope_id);
     }
 
+    #[inline]
     fn visit_ts_interface_declaration(&mut self, it: &TSInterfaceDeclaration<'a>) {
-        self.scope_ids.push(it.scope_id.get().unwrap());
+        self.add_scope(&it.scope_id);
     }
 
+    #[inline]
     fn visit_ts_mapped_type(&mut self, it: &TSMappedType<'a>) {
-        self.scope_ids.push(it.scope_id.get().unwrap());
+        self.add_scope(&it.scope_id);
     }
 
+    #[inline]
     fn visit_ts_conditional_type(&mut self, it: &TSConditionalType<'a>) {
-        self.scope_ids.push(it.scope_id.get().unwrap());
+        self.add_scope(&it.scope_id);
     }
 }