+++ crates/oxc_wasm/src/lib.rs
@@ -241,6 +241,10 @@ impl Oxc {
         }
 
         if run_options.transform() {
+            let (symbols, scopes) = SemanticBuilder::new(source_text, source_type)
+                .build(program)
+                .semantic
+                .into_symbol_table_and_scope_tree();
             let options = TransformOptions::default();
             let result = Transformer::new(
                 &allocator,
@@ -250,7 +254,7 @@ impl Oxc {
                 ret.trivias.clone(),
                 options,
             )
-            .build(program);
+            .build_with_symbols_and_scopes(symbols, scopes, program);
             if !result.errors.is_empty() {
                 let errors = result.errors.into_iter().map(Error::from).collect::<Vec<_>>();
                 self.save_diagnostics(errors);
@@ -309,7 +313,7 @@ impl Oxc {
             semantic: &Semantic,
             scope_text: &mut String,
             depth: usize,
-            scope_ids: &Vec<ScopeId>,
+            scope_ids: &[ScopeId],
         ) {
             let space = " ".repeat(depth * 2);
 
@@ -332,15 +336,13 @@ impl Oxc {
                     scope_text.push_str(&format!("\n{binding_space}}}\n"));
                 }
 
-                if let Some(next_scope_ids) = next_scope_ids {
-                    write_scope_text(semantic, scope_text, depth + 1, next_scope_ids);
-                }
+                write_scope_text(semantic, scope_text, depth + 1, next_scope_ids);
                 scope_text.push_str(&format!("{space}}}\n"));
             }
         }
 
         let mut scope_text = String::default();
-        write_scope_text(semantic, &mut scope_text, 0, &vec![semantic.scopes().root_scope_id()]);
+        write_scope_text(semantic, &mut scope_text, 0, &[semantic.scopes().root_scope_id()]);
         scope_text
     }