+++ tasks/transform_conformance/src/test_case.rs
@@ -5,11 +5,11 @@ use std::{
 
 use oxc::allocator::Allocator;
 use oxc::codegen::CodeGenerator;
-use oxc::diagnostics::{Error, OxcDiagnostic};
+use oxc::diagnostics::{Error, NamedSource, OxcDiagnostic};
 use oxc::parser::Parser;
 use oxc::span::{SourceType, VALID_EXTENSIONS};
 use oxc::transformer::{BabelOptions, TransformOptions};
-use oxc_tasks_common::{normalize_path, print_diff_in_terminal};
+use oxc_tasks_common::{normalize_path, print_diff_in_terminal, project_root};
 
 use crate::{
     constants::{PLUGINS_NOT_SUPPORTED_YET, SKIP_TESTS},
@@ -158,11 +158,11 @@ pub trait TestCase {
         false
     }
 
-    fn transform(&self, path: &Path) -> Result<String, Vec<OxcDiagnostic>> {
+    fn transform(&self, path: &Path, filtered: bool) -> Result<Driver, OxcDiagnostic> {
         let transform_options = match self.transform_options() {
             Ok(transform_options) => transform_options,
             Err(json_err) => {
-                return Err(vec![OxcDiagnostic::error(format!("{json_err:?}"))]);
+                return Err(OxcDiagnostic::error(format!("{json_err:?}")));
             }
         };
 
@@ -178,7 +178,12 @@ pub trait TestCase {
             source_type = source_type.with_typescript(true);
         }
 
-        Driver::new(transform_options.clone()).execute(&source_text, source_type, path)
+        let driver = Driver::new(filtered, transform_options.clone()).execute(
+            &source_text,
+            source_type,
+            path,
+        );
+        Ok(driver)
     }
 }
 
@@ -249,6 +254,7 @@ impl TestCase for ConformanceTestCase {
             println!("output_path: {output_path:?}");
         }
 
+        let project_root = project_root();
         let mut transformed_code = String::new();
         let mut actual_errors = String::new();
         let mut transform_options = None;
@@ -256,18 +262,21 @@ impl TestCase for ConformanceTestCase {
         match self.transform_options() {
             Ok(options) => {
                 transform_options.replace(options.clone());
-                match Driver::new(options.clone()).execute(&input, source_type, &self.path) {
-                    Ok(printed) => {
-                        transformed_code = printed;
-                    }
-                    Err(errors) => {
-                        let error = errors
-                            .into_iter()
-                            .map(|err| err.to_string())
-                            .collect::<Vec<_>>()
-                            .join("\n");
-                        actual_errors = get_babel_error(&error);
-                    }
+                let mut driver =
+                    Driver::new(filtered, options.clone()).execute(&input, source_type, &self.path);
+                transformed_code = driver.printed();
+                let errors = driver.errors();
+                if !errors.is_empty() {
+                    let source = NamedSource::new(
+                        self.path.strip_prefix(project_root).unwrap().to_string_lossy(),
+                        input.to_string(),
+                    );
+                    let error = errors
+                        .into_iter()
+                        .map(|err| format!("{:?}", err.with_source_code(source.clone())))
+                        .collect::<Vec<_>>()
+                        .join("\n");
+                    actual_errors = get_babel_error(&error);
                 }
             }
             Err(json_err) => {
@@ -396,14 +405,11 @@ impl TestCase for ExecTestCase {
             println!("Input:\n{}\n", fs::read_to_string(&self.path).unwrap());
         }
 
-        let result = match self.transform(&self.path) {
-            Ok(result) => result,
+        let result = match self.transform(&self.path, filtered) {
+            Ok(mut driver) => driver.printed(),
             Err(error) => {
                 if filtered {
-                    println!(
-                        "Transform Errors:\n{}\n",
-                        error.iter().map(ToString::to_string).collect::<Vec<_>>().join("\n")
-                    );
+                    println!("Transform Errors:\n{error:?}\n",);
                 }
                 return;
             }